moving samplers to own header
